feat: solve esm.mjs
